Project FW7 Phase 1.
This is a first crude attempt to decouple this horrible monstrosity called "ui.GamePage"
gamePage is now split in ui.GamePage and ui.UISystem.

Some of the UI-related logic was moved there, ideally managers should not care about updating UI.
The next logical step would be to split subclasses of tab.Manager to modelManager and uiManager and store a list of uiManager's into UISystem instance for better portability.