Introduce navigationDiplomacy for sub 1 year trading
Before this patch with Diplomacy purchased trading can only occur after year 1.  In late game short runs this means no trading can occur before 13 minutes and no matter what strategy the player chooses they still have to wait and so the play space they can choose their strategy is limited.

This addition opens all new problems for the players, letting the player decide if they should start trade earlier or if they should concentrate on other aspects or bounce between the two.  Trading could be worth it for Blueprints or iron or other resources v.s. other means or maybe not.  The number of choices they can make grows significantly.

Elsewhere in the game when the player is waiting for something they are able to effect the outcome, but with diplomacy the player just has to wait a year, period.  There isn't anything they can do, nothing they can build to make that go faster.  This addition allows the player to choose to spend their time to build a ship to unlock trade sooner, but the longer they wait to build a ship the less value the "bonus".

To buy navigationDiplomacy you have to have first bought Diplomacy making navigationDiplomacy a simply slight better version of Diplomacy that isn't worth the cost until much later in the game.